Added some more TODOS and started writing more on what to do next.
It would be good if the GLRParser got support for the new token system
asap but that might take some time. Some ideas of how the compiler
layout should be has been fixed and written to the Compiler file. More
grammars might be added soon but the most important step is probably to
generate a working AST with the new token system.